Moves lovell utils to drupal directory
